Combine Configurator with SettingsListener
